+++ contracts/.gitignore
@@ -1,4 +1,2 @@
 /node_modules
 .npmrc
-package-lock.json
-yarn.lock
