+++ contracts/test/index.js
