+++ network/ca/docker-compose.yaml
@@ -11,6 +11,9 @@ services:
       - POSTGRES_DB=fabric_ca
     volumes:
       - postgres_data:/var/lib/postgresql/data
+    networks:
+      front_default:
+        ipv4_address: 172.28.0.2
 
   ca:
     image: hyperledger/fabric-ca:1.4
@@ -24,6 +27,13 @@ services:
     ports:
       - 7054:7054
     command: sh -c "sleep 5 && fabric-ca-server start"
+    networks:
+      front_default:
+        ipv4_address: 172.28.0.3
+
 
 volumes:
   postgres_data:
+networks:
+  front_default:
+    external: true
