+++ network/peer/data/core.yaml
@@ -30,7 +30,7 @@ peer:
     # in the same organization. For peers in other organization, see
     # gossip.externalEndpoint for more info.
     # When used as CLI config, this means the peer's endpoint to interact with
-    address: 165.22.77.44:7051
+    address: 172.28.0.4:7051
 
     # Whether the Peer should programmatically determine its address
     # This case is useful for docker containers.
